Move the top-level PHP files to the src directory
