+++ src/parsely-settings.php
